Alien commander
Fleshes out alien commander + takes the 
lerk viewbobbing from Nin's bots and implements it here. This is done 
instead of raytracing when not in fights, so lerks sometimes get stuck 
but it's very rare and this method uses no raytracing when not in combat 
for lerks.

Alien commander does all the things you'd expect him to do EXCEPT 
contaminate